Add files via upload
this variation of the BAT file is to revert the windows registry key value back to it's original state then uses a loop to directly install the fonts in the same folder the script is running from directly into the Winfows\Fonts folder and register the font names in the Windows registry. no reboot needed to see the result working.